pyproject.toml
[project]
name = "browser-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "browser-use>=0.11.3",
    "python-dotenv>=1.2.1",
]
